fix: spread first load sightings
